pyproject.toml
[project]
name = "prompt-refiner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.76.0",
    "python-fasthtml>=0.12.25",
]
